pyproject.toml
[project]
name = "nitro"
version = "0.1"
description = ""
readme = "README.md"
authors = [{name = ""}]
license = "Apache-2.0"
requires-python = ">=3.8"
classifiers = []
dependencies = [
    "click",
    "datasets>=3.1.0",
    "pandas>=2.0.3",
    "pyyaml>=6.0.2",
    "rich>=14.1.0",
    "ruff>=0.12.5",
    "scikit-learn>=1.3.2",
    "local @ /Users/lliang/projects/nitro/nitro/local"
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

# [tool.uv.workspace]
# members = [
#     "local/local",
#     "local",
# ]

[project.scripts]
nitro = "nitro.cli:cli"

[project.optional-dependencies]
test = ["pytest"]
